pyproject.toml
[project]
name = "mudul-8-exam"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
]
